Move RichHighlighter to rich module
