+++ plextraktsync/rich/RichHighlighter.py
